requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
huggingface-hub==0.17.0
requests==2.31.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.1
sqlalchemy==2.0.23
